Add files via upload
“Himmelblau function”
The objective is to minimize where −6 ≤ x ≤ 6 and −6 ≤ y ≤ 6. One of the global minima lies at (3.0000, 2.0000) where z = 0.0000.
The objective is to minimize z where z = (x^2 + y -11)^2 + (x + y^2 - 7)^2

A simple ES will be coded to solve the problem. To accomplish this, the problem will be encoded as two real variables, values for µ and l will be selected, a standard deviation (the same for both variables) will be chosen and adjusted using the 1/5 success rule, an initial population will be randomly selected, and a stopping criterion will be established.